@@ -1,69 +1,69 @@
 {
   "name": "whatsapp-client-angularcli-material",
   "version": "0.0.0",
-  "license": "MIT",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build --prod",
+    "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "generator": "gql-gen --url http://localhost:3000/graphql --template ts --out ./src/types.d.ts \"./src/graphql/**/*.ts\""
+    "generator": "gql-gen --schema http://localhost:3000/graphql --template ts --out ./src/types.d.ts \"./src/graphql/**/*.ts\""
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.5",
-    "@angular/cdk": "5.2.1",
-    "@angular/common": "5.2.5",
-    "@angular/compiler": "5.2.5",
-    "@angular/core": "5.2.5",
-    "@angular/flex-layout": "2.0.0-beta.12",
-    "@angular/forms": "5.2.5",
-    "@angular/http": "5.2.5",
-    "@angular/material": "5.2.1",
-    "@angular/platform-browser": "5.2.5",
-    "@angular/platform-browser-dynamic": "5.2.5",
-    "@angular/router": "5.2.5",
-    "ajv": "6.1.1",
-    "apollo-angular": "1.0.1",
-    "apollo-angular-link-http": "1.0.2-beta.0",
-    "apollo-cache-inmemory": "1.1.9",
-    "apollo-client": "2.2.5",
-    "apollo-link": "1.1.0",
-    "apollo-link-ws": "1.0.5",
-    "apollo-utilities": "1.0.8",
-    "core-js": "2.5.3",
-    "graphql": "0.12.3",
-    "graphql-tag": "2.7.3",
+    "@angular/animations": "6.0.0",
+    "@angular/cdk": "6.0.1",
+    "@angular/common": "6.0.0",
+    "@angular/compiler": "6.0.0",
+    "@angular/core": "6.0.0",
+    "@angular/flex-layout": "6.0.0-beta.15",
+    "@angular/forms": "6.0.0",
+    "@angular/http": "6.0.0",
+    "@angular/material": "6.0.1",
+    "@angular/platform-browser": "6.0.0",
+    "@angular/platform-browser-dynamic": "6.0.0",
+    "@angular/router": "6.0.0",
+    "apollo-angular": "1.1.0-rc.0",
+    "apollo-angular-link-http": "1.1.0-rc.0",
+    "apollo-cache-inmemory": "1.2.1",
+    "apollo-client": "2.3.1",
+    "apollo-link": "1.2.2",
+    "apollo-link-ws": "1.0.8",
+    "apollo-utilities": "1.0.12",
+    "core-js": "2.5.6",
+    "graphql": "0.13.2",
+    "graphql-tag": "2.9.2",
     "hammerjs": "2.0.8",
-    "moment": "2.20.1",
-    "ng2-truncate": "1.3.11",
-    "ngx-selectable-list": "1.1.0",
-    "rxjs": "5.5.6",
-    "subscriptions-transport-ws": "0.9.5",
-    "zone.js": "0.8.20"
+    "moment": "2.22.1",
+    "ng2-truncate": "1.3.17",
+    "ngx-selectable-list": "^1.2.1",
+    "rxjs": "6.1.0",
+    "subscriptions-transport-ws": "0.9.9",
+    "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.0",
-    "@angular/compiler-cli": "5.2.5",
-    "@angular/language-service": "5.2.5",
-    "@types/graphql": "0.12.4",
-    "@types/jasmine": "2.8.6",
+    "@angular-devkit/build-angular": "0.6.0",
+    "@angular/cli": "6.0.0",
+    "@angular/compiler-cli": "6.0.0",
+    "@angular/language-service": "6.0.0",
+    "@types/graphql": "0.13.0",
+    "@types/jasmine": "2.8.7",
     "@types/jasminewd2": "2.0.3",
-    "@types/node": "6.0.101",
-    "codelyzer": "4.1.0",
-    "graphql-code-generator": "0.8.14",
-    "jasmine-core": "2.8.0",
+    "@types/node": "8.9.5",
+    "codelyzer": "4.2.1",
+    "graphql-code-generator": "0.9.1",
+    "graphql-codegen-typescript-template": "0.9.1",
+    "jasmine-core": "2.99.1",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "2.0.0",
+    "karma": "1.7.1",
     "karma-chrome-launcher": "2.2.0",
-    "karma-coverage-istanbul-reporter": "1.4.1",
-    "karma-jasmine": "1.1.1",
+    "karma-coverage-istanbul-reporter": "1.4.2",
+    "karma-jasmine": "1.1.2",
     "karma-jasmine-html-reporter": "0.2.2",
-    "protractor": "5.1.2",
-    "ts-node": "4.1.0",
+    "protractor": "5.3.1",
+    "ts-node": "5.0.1",
     "tslint": "5.9.1",
-    "typescript": "2.5.3"
+    "typescript": "2.7.2"
   }
 }
