@@ -1,10 +1,10 @@
-import {ApolloQueryResult, MutationOptions, WatchQueryOptions} from 'apollo-client';
+import {ApolloQueryResult, FetchMoreOptions, FetchMoreQueryOptions, MutationOptions, WatchQueryOptions} from 'apollo-client';
 import {concat, map, share, switchMap} from 'rxjs/operators';
 import {Apollo, QueryRef} from 'apollo-angular';
 import {Injectable} from '@angular/core';
 import {getChatsQuery} from '../../graphql/getChats.query';
 import {
-  AddChat, AddGroup, AddMessage, GetChat, GetChats, GetUsers, MessageAdded, RemoveAllMessages, RemoveChat,
+  AddChat, AddGroup, AddMessage, GetChat, GetChats, GetUsers, MessageAdded, MoreMessages, RemoveAllMessages, RemoveChat,
   RemoveMessages
 } from '../../types';
 import {getChatQuery} from '../../graphql/getChat.query';
@@ -14,20 +14,20 @@ import {DocumentNode} from 'graphql';
 import {removeAllMessagesMutation} from '../../graphql/removeAllMessages.mutation';
 import {removeMessagesMutation} from '../../graphql/removeMessages.mutation';
 import {getUsersQuery} from '../../graphql/getUsers.query';
-import {Observable} from 'rxjs/Observable';
+import {Observable, AsyncSubject, of} from 'rxjs';
 import {addChatMutation} from '../../graphql/addChat.mutation';
 import {addGroupMutation} from '../../graphql/addGroup.mutation';
 import * as moment from 'moment';
-import {AsyncSubject} from 'rxjs/AsyncSubject';
-import {of} from 'rxjs/observable/of';
 import {FetchResult} from 'apollo-link';
 import {LoginService} from '../login/services/login.service';
 import {chatAddedSubscription} from '../../graphql/chatAdded.subscription';
 import {messageAddedSubscription} from '../../graphql/messageAdded.subscription';
+import {moreMessagesQuery} from '../../graphql/moreMessages.query';
 
 @Injectable()
 export class ChatsService {
-  messagesAmount = 3;
+  chatsMessagesAmount = 2;
+  chatMessagesAmount = 5;
   getChatsWq: QueryRef<GetChats.Query>;
   chats$: Observable<GetChats.Chats[]>;
   chats: GetChats.Chats[];
@@ -39,12 +39,15 @@ export class ChatsService {
     this.getChatsWq = this.apollo.watchQuery<GetChats.Query>(<WatchQueryOptions>{
       query: getChatsQuery,
       variables: {
-        amount: this.messagesAmount,
+        amount: this.chatsMessagesAmount,
       },
     });
 
     this.getChatsWq.subscribeToMore({
       document: chatAddedSubscription,
+      variables: {
+        amount: this.chatsMessagesAmount,
+      },
       updateQuery: (prev: GetChats.Query, { subscriptionData }) => {
         if (!subscriptionData.data) {
           return prev;
@@ -73,20 +76,25 @@ export class ChatsService {
           const {chat}: GetChat.Query = this.apollo.getClient().readQuery({
             query: getChatQuery, variables: {
               chatId: newMessage.chat.id,
+              amount: this.chatMessagesAmount,
             }
           });
 
           // Add our message from the mutation to the end.
-          chat.messages.push(newMessage);
+          chat.messageFeed.messages.push(newMessage);
           // Write our data back to the cache.
-          this.apollo.getClient().writeQuery({ query: getChatQuery, data: {chat} });
+          this.apollo.getClient().writeQuery({
+            query: getChatQuery,
+            variables: {chatId: newMessage.chat.id, amount: this.chatMessagesAmount},
+            data: {chat} });
         } catch {
           console.error('The chat we received an update for does not exist in the store');
         }
 
         return Object.assign({}, prev, {
-          chats: [...prev.chats.map(_chat =>
-            _chat.id === newMessage.chat.id ? {..._chat, messages: [..._chat.messages, newMessage]} : _chat)]
+          chats: [...prev.chats.map(chat => chat.id === newMessage.chat.id ? {
+            ...chat, messageFeed: {...chat.messageFeed, messages: [...chat.messageFeed.messages, newMessage]}
+          } : chat)]
         });
       }
     });
@@ -105,7 +113,7 @@ export class ChatsService {
     return {query: this.getChatsWq, chats$: this.chats$};
   }
 
-  getChat(chatId: string, oui?: boolean) {
+  getChat(chatId: string, oui?: boolean, amount?: number) {
     const _chat = this.chats && this.chats.find(chat => chat.id === chatId) || {
       id: chatId,
       name: '',
@@ -113,7 +121,11 @@ export class ChatsService {
       allTimeMembers: [],
       unreadMessages: 0,
       isGroup: false,
-      messages: [],
+      messageFeed: {
+        hasNextPage: false,
+        cursor: null,
+        messages: [],
+      },
     };
     const chat$FromCache = of<GetChat.Chat>(_chat);
 
@@ -122,6 +134,7 @@ export class ChatsService {
         query: getChatQuery,
         variables: {
           chatId: id,
+          amount: this.chatMessagesAmount,
         }
       });
     };
@@ -154,6 +167,34 @@ export class ChatsService {
     return {query$: this.getChatWqSubject.asObservable(), chat$};
   }
 
+  moreMessages(query: QueryRef<GetChat.Query>, chatId: string) {
+    const {data: {chat: {messageFeed}}} = query.getLastResult();
+    if (messageFeed.hasNextPage) {
+      query.fetchMore({
+        query: moreMessagesQuery,
+        variables: {
+          chatId,
+          amount: this.chatMessagesAmount,
+          before: messageFeed.cursor,
+        },
+        updateQuery: (previousResult: GetChat.Query, { fetchMoreResult }) => {
+          return {
+            chat: {
+              ...previousResult.chat,
+              messageFeed: {
+                ...fetchMoreResult.chat.messageFeed,
+                messages: [
+                  ...fetchMoreResult.chat.messageFeed.messages,
+                  ...previousResult.chat.messageFeed.messages,
+                ],
+              },
+            },
+          };
+        },
+      });
+    }
+  }
+
   addMessage(chatId: string, content: string) {
     return this.apollo.mutate(<MutationOptions>{
       mutation: addMessageMutation,
@@ -166,14 +207,17 @@ export class ChatsService {
         addMessage: {
           id: ChatsService.getRandomId(),
           __typename: 'Message',
-          senderId: this.loginService.getUser().id,
+          chat: {
+            id: chatId,
+            __typename: 'Chat',
+          },
           sender: {
             id: this.loginService.getUser().id,
             __typename: 'User',
             name: this.loginService.getUser().name,
           },
           content,
-          createdAt: moment().unix(),
+          createdAt: new Date(),
           type: 0,
           recipients: [],
           ownership: true,
@@ -186,12 +230,13 @@ export class ChatsService {
           const {chat}: GetChat.Query = store.readQuery({
             query: getChatQuery, variables: {
               chatId,
+              amount: this.chatMessagesAmount,
             }
           });
           // Add our message from the mutation to the end.
-          chat.messages.push(addMessage);
+          chat.messageFeed.messages.push(addMessage);
           // Write our data back to the cache.
-          store.writeQuery({ query: getChatQuery, data: {chat} });
+          store.writeQuery({ query: getChatQuery, variables: {chatId, amount: this.chatMessagesAmount}, data: {chat} });
         }
         // Update last message cache
         {
@@ -199,16 +244,16 @@ export class ChatsService {
           const {chats}: GetChats.Query = store.readQuery({
             query: getChatsQuery,
             variables: <GetChats.Variables>{
-              amount: this.messagesAmount,
+              amount: this.chatsMessagesAmount,
             },
           });
           // Add our comment from the mutation to the end.
-          chats.find(chat => chat.id === chatId).messages.push(addMessage);
+          chats.find(chat => chat.id === chatId).messageFeed.messages.push(addMessage);
           // Write our data back to the cache.
           store.writeQuery({
             query: getChatsQuery,
             variables: <GetChats.Variables>{
-              amount: this.messagesAmount,
+              amount: this.chatsMessagesAmount,
             },
             data: {
               chats,
@@ -234,7 +279,7 @@ export class ChatsService {
         const {chats}: GetChats.Query = store.readQuery({
           query: getChatsQuery,
           variables: <GetChats.Variables>{
-            amount: this.messagesAmount,
+            amount: this.chatsMessagesAmount,
           },
         });
         // Remove the chat (mutable)
@@ -247,7 +292,7 @@ export class ChatsService {
         store.writeQuery({
           query: getChatsQuery,
           variables: <GetChats.Variables>{
-            amount: this.messagesAmount,
+            amount: this.chatsMessagesAmount,
           },
           data: {
             chats,
@@ -286,18 +331,19 @@ export class ChatsService {
           const {chat}: GetChat.Query = store.readQuery({
             query: getChatQuery, variables: {
               chatId,
+              amount: this.chatMessagesAmount,
             }
           });
           // Remove the messages (mutable)
           removeMessages.forEach(messageId => {
-            for (const index of chat.messages.keys()) {
-              if (chat.messages[index].id === messageId) {
-                chat.messages.splice(index, 1);
+            for (const index of chat.messageFeed.messages.keys()) {
+              if (chat.messageFeed.messages[index].id === messageId) {
+                chat.messageFeed.messages.splice(index, 1);
               }
             }
           });
           // Write our data back to the cache.
-          store.writeQuery({ query: getChatQuery, data: {chat} });
+          store.writeQuery({ query: getChatQuery, variables: {chatId, amount: this.chatMessagesAmount}, data: {chat} });
         }
         // Update last message cache
         {
@@ -305,18 +351,18 @@ export class ChatsService {
           const {chats}: GetChats.Query = store.readQuery({
             query: getChatsQuery,
             variables: <GetChats.Variables>{
-              amount: this.messagesAmount,
+              amount: this.chatsMessagesAmount,
             },
           });
           // Fix last message
-          chats.find(chat => chat.id === chatId).messages = messages
+          chats.find(chat => chat.id === chatId).messageFeed.messages = messages
             .filter(message => !ids.includes(message.id))
             .sort((a, b) => Number(b.createdAt) - Number(a.createdAt)) || [];
           // Write our data back to the cache.
           store.writeQuery({
             query: getChatsQuery,
             variables: <GetChats.Variables>{
-              amount: this.messagesAmount,
+              amount: this.chatsMessagesAmount,
             },
             data: {
               chats,
@@ -371,7 +417,12 @@ export class ChatsService {
             }
           ],
           unreadMessages: 0,
-          messages: [],
+          messageFeed: {
+            __typename: 'MessageFeed',
+            hasNextPage: false,
+            cursor: null,
+            messages: [],
+          },
           isGroup: false,
         },
       },
@@ -380,7 +431,7 @@ export class ChatsService {
         const {chats}: GetChats.Query = store.readQuery({
           query: getChatsQuery,
           variables: <GetChats.Variables>{
-            amount: this.messagesAmount,
+            amount: this.chatsMessagesAmount,
           },
         });
         // Add our comment from the mutation to the end.
@@ -389,7 +440,7 @@ export class ChatsService {
         store.writeQuery({
           query: getChatsQuery,
           variables: <GetChats.Variables>{
-            amount: this.messagesAmount,
+            amount: this.chatsMessagesAmount,
           },
           data: {
             chats,
@@ -423,7 +474,12 @@ export class ChatsService {
             ...recipientIds.map(id => ({id, __typename: 'User'})),
           ],
           unreadMessages: 0,
-          messages: [],
+          messageFeed: {
+            __typename: 'MessageFeed',
+            hasNextPage: false,
+            cursor: null,
+            messages: [],
+          },
           isGroup: true,
         },
       },
@@ -432,7 +488,7 @@ export class ChatsService {
         const {chats}: GetChats.Query = store.readQuery({
           query: getChatsQuery,
           variables: <GetChats.Variables>{
-            amount: this.messagesAmount,
+            amount: this.chatsMessagesAmount,
           },
         });
         // Add our comment from the mutation to the end.
@@ -441,7 +497,7 @@ export class ChatsService {
         store.writeQuery({
           query: getChatsQuery,
           variables: <GetChats.Variables>{
-            amount: this.messagesAmount,
+            amount: this.chatsMessagesAmount,
           },
           data: {
             chats,
