@@ -11,7 +11,7 @@ import {MessagesListComponent} from './components/messages-list/messages-list.co
 import {MessageItemComponent} from './components/message-item/message-item.component';
 import {NewMessageComponent} from './components/new-message/new-message.component';
 import {SharedModule} from '../shared/shared.module';
-import {SelectableListModule} from 'ngx-selectable-list';
+import {NgxSelectableListModule} from 'ngx-selectable-list';
 import {AuthGuard} from '../login/services/auth.guard';
 
 const routes: Routes = [
@@ -46,7 +46,7 @@ const routes: Routes = [
     FormsModule,
     // Feature modules
     SharedModule,
-    SelectableListModule,
+    NgxSelectableListModule,
   ],
   providers: [
     ChatsService,
