@@ -3,11 +3,15 @@ import {fragments} from './fragment';
 
 // We use the gql tag to parse our query string into a query document
 export const chatAddedSubscription = gql`
-  subscription chatAdded {
+  subscription chatAdded($amount: Int!) {
     chatAdded {
       ...ChatWithoutMessages
-      messages {
-        ...Message
+      messageFeed(amount: $amount) {
+        hasNextPage,
+        cursor,
+        messages {
+          ...Message
+        }
       }
     }
   }
