@@ -3,11 +3,15 @@ import {fragments} from './fragment';
 
 // We use the gql tag to parse our query string into a query document
 export const getChatQuery = gql`
-  query GetChat($chatId: ID!) {
+  query GetChat($chatId: ID!, $amount: Int!) {
     chat(chatId: $chatId) {
       ...ChatWithoutMessages
-      messages {
-        ...Message
+      messageFeed(amount: $amount) {
+        hasNextPage,
+        cursor,
+        messages {
+          ...Message
+        }
       }
     }
   }
