@@ -3,11 +3,15 @@ import {fragments} from './fragment';
 
 // We use the gql tag to parse our query string into a query document
 export const getChatsQuery = gql`
-  query GetChats($amount: Int) {
+  query GetChats($amount: Int!) {
     chats {
       ...ChatWithoutMessages
-      messages(amount: $amount) {
-        ...Message
+      messageFeed(amount: $amount) {
+        hasNextPage,
+        cursor,
+        messages {
+          ...Message
+        }
       }
     }
   }
